vqa-ai-project/requirements.txt
flask==3.0.3
google-generativeai==0.7.2
Pillow==10.4.0
python-dotenv==1.0.1
deep-translator==1.11.4
werkzeug==3.0.4
gunicorn==22.0.0
